Basic Dockerfile demo
This commit adds a very basic Dockerfile to let you run the scribe
application in a container. This is not recommended in any way as we
end up with a single layer with the whole application so every time
something changes, the whole layer has to be rebuilt again.

From the root directory of our project, we can build this image provided
that the Docker daemon is available:

docker build -t scribe-custom:0.0.1-SNAPSHOT .

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe-custom:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe-custom:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe-custom:0.0.1-SNAPSHOT