Application layers Dockerfile demo
This commit improves our Dockerfile to create separate layers for the
libraries and the application classes. We first extract the content of
the jar and we copy differents assets in different layers.

Our entrypoint has to build the classpath and has to know the main class
which is a bit annoying if you want to keep the flexibility of using
Spring Boot's launcher.

From the root directory of our project, we can build this image provided
that the Docker daemon is available:

docker build -t scribe-custom:0.0.1-SNAPSHOT .

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe-custom:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe-custom:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe-custom:0.0.1-SNAPSHOT