Custom Application layers Dockerfile with Spring Boot demo
Spring Boot allows you to define the layers of your application in a
more fine-grained manner. For the sake of the example, let's separate
the webjars of our application in a separate layer as they change more
often than other dependencies. And we don't use snapshots at the moment,
so we can remove that dedicated layer.

This commits defines this extra layer for both Maven and Gradle. Maven
uses a separate configuration file in XML format while Gradle uses a
dedicated DSL.

Once we're rebuilt the archive, we can run the jar again to list the
updated layers:

java -Djarmode=layertools -jar scribe/target/scribe-*.jar list

This now yields the following layers:

* dependencies
* spring-boot-loader
* webjars
* application

This commit also updates the `Dockerfile` to copy the webjars layer in
the generated image. This is important as keeping the previous
definition won't work since the "snapshots-depdencies" layer is gone and
the resulting image when fixing this issue would miss all webjars.

From the root directory of our project, we can build this image provided
that the Docker daemon is available:

docker build -t scribe-custom:0.0.1-SNAPSHOT .

We can run the app in foreground as follows:

docker run --rm -p 8080:8080 scribe-custom:0.0.1-SNAPSHOT

If we want to set an environment variable to customize the application's
configuration, we can use the `-e` flag:

docker run --rm -p 8080:8080 -e CUSTOM_TEST_PROPERTY=cmdline\
    scribe-custom:0.0.1-SNAPSHOT

We can finally inspect the generated image with `dive` to confirm the
application is built out of a single layer:

dive scribe-custom:0.0.1-SNAPSHOT