Liveness and Readiness probe demo
Spring Boot enforces best practices when running on Kubernetes without
enforcing a hard dependency on Kubernetes. Some of those best practices
are also applicable to other environments.

In this commit, we enable the health probes via configuration. This
exposes two additional healh groups called `liveness` and `readiness`.

When running on Kubernetes, this property is not necessary as Spring
Boot enables those probes automatically.